+++ packages/rs-drive/src/drive/batch/action_convert_to_operations/identity/mod.rs
@@ -1,5 +1,5 @@
 mod identity_create_transition;
-mod identity_credit_transfer;
+mod identity_credit_transfer_transition;
 mod identity_credit_withdrawal_transition;
 mod identity_top_up_transition;
 mod identity_update_transition;
