+++ packages/rs-drive/src/drive/batch/mod.rs
@@ -6,8 +6,8 @@
 /// Operation module
 pub mod drive_op_batch;
 // TODO: Must crate only but we need to remove of use it first
+pub mod action_convert_to_operations;
 pub mod grovedb_op_batch;
-pub mod transitions;
 
 pub use drive_op_batch::DataContractOperationType;
 pub use drive_op_batch::DocumentOperationType;
