+++ packages/rs-platform-version/src/version/v1.rs
@@ -63,16 +63,17 @@ use crate::version::drive_versions::{
     DriveIdentityWithdrawalTransactionQueueMethodVersions, DriveInitializationMethodVersions,
     DriveMethodVersions, DriveOperationsMethodVersion, DrivePlatformStateMethodVersions,
     DrivePlatformSystemMethodVersions, DrivePrefundedSpecializedMethodVersions,
-    DriveProtocolUpgradeVersions, DriveProveMethodVersions, DriveStateTransitionMethodVersions,
-    DriveStateTransitionOperationMethodVersions, DriveStructureVersion,
-    DriveSystemEstimationCostsMethodVersions, DriveVerifyContractMethodVersions,
-    DriveVerifyDocumentMethodVersions, DriveVerifyIdentityMethodVersions,
-    DriveVerifyMethodVersions, DriveVerifySingleDocumentMethodVersions,
-    DriveVerifyStateTransitionMethodVersions, DriveVerifySystemMethodVersions,
-    DriveVerifyVoteMethodVersions, DriveVersion, DriveVoteCleanupMethodVersions,
-    DriveVoteContestedResourceInsertMethodVersions, DriveVoteFetchMethodVersions,
-    DriveVoteInsertMethodVersions, DriveVoteMethodVersions, DriveVoteSetupMethodVersions,
-    DriveVoteStorageFormMethodVersions,
+    DriveProtocolUpgradeVersions, DriveProveMethodVersions,
+    DriveStateTransitionActionConvertToHighLevelOperationsMethodVersions,
+    DriveStateTransitionMethodVersions, DriveStateTransitionOperationMethodVersions,
+    DriveStructureVersion, DriveSystemEstimationCostsMethodVersions,
+    DriveVerifyContractMethodVersions, DriveVerifyDocumentMethodVersions,
+    DriveVerifyIdentityMethodVersions, DriveVerifyMethodVersions,
+    DriveVerifySingleDocumentMethodVersions, DriveVerifyStateTransitionMethodVersions,
+    DriveVerifySystemMethodVersions, DriveVerifyVoteMethodVersions, DriveVersion,
+    DriveVoteCleanupMethodVersions, DriveVoteContestedResourceInsertMethodVersions,
+    DriveVoteFetchMethodVersions, DriveVoteInsertMethodVersions, DriveVoteMethodVersions,
+    DriveVoteSetupMethodVersions, DriveVoteStorageFormMethodVersions,
 };
 use crate::version::fee::v1::FEE_VERSION1;
 use crate::version::protocol_version::{FeatureVersionBounds, PlatformVersion};
@@ -492,6 +493,27 @@ pub const PLATFORM_V1: PlatformVersion = PlatformVersion {
                         finalization_tasks: 0,
                     },
                 },
+                convert_to_high_level_operations:
+                    DriveStateTransitionActionConvertToHighLevelOperationsMethodVersions {
+                        data_contract_create_transition: 0,
+                        data_contract_update_transition: 0,
+                        document_create_transition: 0,
+                        document_delete_transition: 0,
+                        document_purchase_transition: 0,
+                        document_replace_transition: 0,
+                        document_transfer_transition: 0,
+                        document_update_price_transition: 0,
+                        documents_batch_transition: 0,
+                        identity_create_transition: 0,
+                        identity_credit_transfer_transition: 0,
+                        identity_credit_withdrawal_transition: 0,
+                        identity_top_up_transition: 0,
+                        identity_update_transition: 0,
+                        masternode_vote_transition: 0,
+                        bump_identity_data_contract_nonce: 0,
+                        bump_identity_nonce: 0,
+                        partially_use_asset_lock: 0,
+                    },
             },
             batch_operations: DriveBatchOperationsMethodVersion {
                 convert_drive_operations_to_grove_operations: 0,
